requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
google-adk>=0.1.0
google-generativeai>=0.5.0
gspread>=6.0.0
google-auth>=2.28.0
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.27.0
pydantic>=2.0.0
reportlab>=4.1.0
pypdf>=4.0.0
python-multipart>=0.0.9
aiofiles>=23.2.0
rich>=13.7.0
